+++ database/migrations/2021_06_15_205945_create_lections_table.php
@@ -15,7 +15,7 @@ class CreateLectionsTable extends Migration
     {
         Schema::create('lections', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('classes_id');
+            $table->unsignedBigInteger('classes_id');
             $table->string('name');
             $table->timestamps();
